Fixes syntax, adds try-catch for business rule triggering
I tried to make this as fill-in-the-blank as possible. I put TODO on things I either left blank or suspect there may be an issue when you're done writing code. It's your job to make sure your class in the model is doing the right thing, and adjusting this view if necessary.